# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Book",
        "description": "desplay books whith gender and author",
        "contact": {
            "email": "adrien.charrier@outlook.fr"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/authors": {
            "get": {
                "tags": [
                    "author"
                ],
                "summary": "Get List Of Authors",
                "description": "Returns all authorss.",
                "operationId": "getAllAuthors",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "make a search in author's name",
                        "required": false
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sort results with the author's name",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "author"
                ],
                "summary": "Store A New Authors",
                "description": "Store A New Authors",
                "operationId": "StoreANewAuthors",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "author name",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/authors/{id}": {
            "get": {
                "tags": [
                    "author"
                ],
                "summary": "Get An Authors",
                "description": "Get An Authors",
                "operationId": "GetAnAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "author id",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "author"
                ],
                "summary": "Delete An Authors",
                "description": "Delete An Authors",
                "operationId": "DeleteAnAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "author id",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "author"
                ],
                "summary": "Update An Authors",
                "description": "Update An Authors",
                "operationId": "UpdateAnAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "author id",
                        "required": true
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "author name",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        []
    ]
}